+++ HISTORY.md
@@ -54,6 +54,7 @@
 
 ### 2023-05-23
 * blocks: implemented postfix functionality for variadic slot labels
+* blocks: implemented label groups for variadic slots
 
 ### 2023-05-22
 * gui: show a login message for student accounts, thanks, Michael!